Added function and button to clear search history and styling to search history
