+++ src/Views/Home/styles.scss
@@ -1,0 +1,63 @@
+@import "src/assets/styles/_variables.scss";
+
+.home-container {
+  .content-wrapper {
+    display: flex;
+    flex-wrap: wrap;
+
+    @media screen and (max-width: $md) {
+      justify-content: center;
+    }
+
+    .search-container {
+      width: 70%;
+    }
+  }
+  .search-history {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    background-color: #fafafa;
+    min-width: 350px;
+    min-height: 300px;
+
+    .empty {
+      margin-top: 20px;
+    }
+
+    h2 {
+      background-color: #f7f1d5;
+      padding: 2vw;
+      margin: 0;
+    }
+
+    ul {
+      list-style: none;
+      padding: 0;
+      margin: 0;
+      li:nth-child(even) {
+        background: #eff0f1;
+      }
+      li {
+        display: flex;
+        justify-content: space-between;
+        padding: 1rem 0;
+        span {
+          margin: 0 1rem;
+        }
+        button {
+          margin-right: 1rem;
+        }
+      }
+    }
+    .clear-history {
+      width: 100%;
+      display: flex;
+      justify-content: flex-end;
+
+      button {
+        margin: 0.75rem;
+      }
+    }
+  }
+}
