+++ src/assets/styles/_variables.scss
@@ -1,0 +1,5 @@
+/* Width breakpoints */
+$xs: 599px;
+$sm: 600px;
+$md: 768px;
+$lg: 1200px;
